// src/serverApi/v3/models/custom-parameter-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CustomParameterResponse
 */
export interface CustomParameterResponse {
    /**
     * 
     * @type {string}
     * @memberof CustomParameterResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof CustomParameterResponse
     */
    defaultValue?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomParameterResponse
     */
    regex?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomParameterResponse
     */
    regexComment?: string;
    /**
     * 
     * @type {string}
     * @memberof CustomParameterResponse
     */
    scope: CustomParameterResponseScope;
    /**
     * 
     * @type {string}
     * @memberof CustomParameterResponse
     */
    location: CustomParameterResponseLocation;
    /**
     * 
     * @type {string}
     * @memberof CustomParameterResponse
     */
    type: CustomParameterResponseType;
    /**
     * 
     * @type {boolean}
     * @memberof CustomParameterResponse
     */
    isOptional: boolean;
}

/**
    * @export
    * @enum {string}
    */
export enum CustomParameterResponseScope {
    GLOBAL = 'global',
    SCHOOL = 'school',
    COURSE = 'course'
}
/**
    * @export
    * @enum {string}
    */
export enum CustomParameterResponseLocation {
    PATH = 'path',
    TOKEN = 'token',
    QUERY = 'query'
}
/**
    * @export
    * @enum {string}
    */
export enum CustomParameterResponseType {
    STRING = 'string',
    NUMBER = 'number',
    BOOLEAN = 'boolean',
    AUTO_COURSEID = 'auto_courseid',
    AUTO_COURSENAME = 'auto_coursename',
    AUTO_SCHOOLID = 'auto_schoolid'
}